redox-endpoint/redox/requirements.txt
cffi==1.13.2
cryptography==2.8
pycparser==2.19
PyJWT==1.7.1
six==1.14.0
